Remove nesting modules
