+++ lib/godmin/resources/resource_service.rb
@@ -1,6 +1,5 @@
 require "godmin/resources/resource_service/batch_actions"
 require "godmin/resources/resource_service/filters"
-require "godmin/resources/resource_service/nesting"
 require "godmin/resources/resource_service/ordering"
 require "godmin/resources/resource_service/pagination"
 require "godmin/resources/resource_service/scopes"
@@ -22,7 +21,6 @@ module Godmin
         @options = options
       end
 
-      # TODO: should this raise its own error?
       def resource_class
         @options[:resource_class] || resource_class_name.constantize
       end
@@ -32,7 +30,11 @@ module Godmin
       end
 
       def resources_relation
-        resource_class.all
+        if options[:resource_parent].present?
+          resource_class.where(options[:resource_parent].class.name.underscore => options[:resource_parent])
+        else
+          resource_class.all
+        end
       end
 
       def resources(params)
